pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pink-elevenlabs"
version = "1.0.0"
description = "Text-to-speech and voice transformation using ElevenLabs API"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.12"
authors = [
    {name = "pinkhairedboy", email = "p.nkhaired@gmail.com"}
]
keywords = ["tts", "text-to-speech", "elevenlabs", "voice-synthesis", "voice-changer"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
]

dependencies = [
    "elevenlabs>=2.21.0",
    "python-dotenv>=1.0.0",
]

[project.urls]
Repository = "https://github.com/pinkhairedboy/pink-elevenlabs"

[project.scripts]
pink-elevenlabs = "pink_elevenlabs.cli.main:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
